This is windows based chatting application for UNCC students
The project aims at developing an instant messenger application for
windows desktops. It allows UNCC students to have the ability to
communicate with one another in an easy manner. In our day-to-day
activities, we come across a number of instances when we need to contact
our fellow UNCC students. Because of unavailability of an instant
messaging application, we are left with only one option that is,
dropping a formal email. Unfortunately, there is one problem with having
an email conversation. At times, you might not get an immediate response
to your email. To solve this problem, we developed a messaging
application that allows students to communicate informally.